meter2mqtt/requirements.txt
requests>=2.25.0
paho-mqtt>=1.5.0
PyYAML>=5.4.0
tenacity>=8.0.0
python-dotenv>=0.19.0 